add to the TileImprovementDB: - a Sound flag, - a Freight flag and - remove the limit of 3 for contruction tiles.
So far only the additional Sound flag is used so far. The limit removement should allow modders to have tile improvements with lots more then 3 construction tiles so that every turn while the tileimp is under construction it changes it appearance.

The Freight flag is now only there and does nothing, but it should later allow to redirect trade routes when roads give such a bonus.