Added CreateBuilding and CreateWonder slic functions.
The difference with the same events is that the functions don't effect stored production in a city. And in the case of the CreateWonder event, doesn't trigger the wonder movie.

And actually I think the CreateWonder event is buggy when generated from slic, because the game freezes. Possibly related to the movie.